added Reinforcement mode

+++ Lingoleap - BE/src/main/java/com/lingo_leap/repository/UserRepository.java
@@ -1,7 +1,5 @@
 package com.lingo_leap.repository;
 
-
-
 import com.lingo_leap.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -9,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByUsername(String username);
 }
